fix(ariaSelectedAttributeRemoval): Refactored code for maintainability and clarity
