--- src/Analysis/SharedUtils/AccessibilityUtils/AntDAccessibilityFixes/Utils/Constants.js
@@ -1,3 +1,0 @@
-const DELAY_TIME = 500;
-
-export { DELAY_TIME };
