+++ src/Analysis/SharedUtils/AccessibilityUtils/AntDAccessibilityFixes/Utils/RemoveDOMAttribute.js
@@ -4,18 +4,15 @@
  */
 
 const RemoveDOMAttribute = (selector, attribute) => {
-    const elements = document.querySelectorAll(selector);
-    console.log('elements',elements, selector)
-    elements?.forEach((element) => {
-      if (element) {
-        console.log('removing element and attribute', element, attribute);
-        element.removeAttribute(attribute);
-      } else {
-        /* eslint-disable no-console */
-        console.error('Unable to find selector in removeDOMAttribute: ', selector);
-        /* eslint-enable no-console */
-      }
-    });
+  const elements = document.querySelectorAll(selector);
+  elements?.forEach((element) => {
+    if (element) {
+      element.removeAttribute(attribute);
+    } else {
+      // eslint-disable-next-line no-console
+      console.error('Unable to find selector in removeDOMAttribute: ', selector);
+    }
+  });
 };
 
 export default RemoveDOMAttribute;
